centralize powershell ceremony to avoid duplication
